backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
openai>=1.40.0
python-dotenv>=1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
